refactor: add integration test helper
