+++ src/test/kotlin/org/fg/ttrpg/CampaignEventResourceIT.kt
@@ -6,34 +6,19 @@ import io.quarkus.test.security.jwt.Claim
 import io.quarkus.test.security.jwt.JwtSecurity
 import io.restassured.RestAssured.given
 import io.restassured.http.ContentType
-import jakarta.inject.Inject
-import jakarta.transaction.Transactional
-import org.fg.ttrpg.account.GM
-import org.fg.ttrpg.account.GMRepository
 import org.fg.ttrpg.campaign.Campaign
-import org.fg.ttrpg.campaign.CampaignRepository
-import org.fg.ttrpg.common.dto.CalendarDTO
 import org.fg.ttrpg.common.dto.CampaignEventOverrideDTO
 import org.fg.ttrpg.common.dto.TimelineEventDTO
 import org.fg.ttrpg.setting.Setting
-import org.fg.ttrpg.setting.SettingRepository
+import org.fg.ttrpg.testutils.IntegrationTestHelper
 import org.hamcrest.CoreMatchers.equalTo
-import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.time.Instant
 import java.util.*
 
 @QuarkusTest
-class CampaignEventResourceIT {
-    @Inject
-    lateinit var gmRepo: GMRepository
-
-    @Inject
-    lateinit var settingRepo: SettingRepository
-
-    @Inject
-    lateinit var campaignRepo: CampaignRepository
+class CampaignEventResourceIT : IntegrationTestHelper() {
 
     val gmId = UUID.fromString("00000000-0000-0000-0000-000000000020")
     lateinit var setting: Setting
@@ -42,66 +27,6 @@ class CampaignEventResourceIT {
     fun setup() {
         createGm(gmId)
         setting = createSetting(UUID.randomUUID(), gmId)
-    }
-
-    @AfterEach
-    fun cleanup() {
-        gmRepo.deleteById(gmId)
-    }
-
-    @Transactional
-    fun createGm(id: UUID) {
-        val gm = GM().apply {
-            this.id = id
-            username = "gm-$id"
-        }
-        gmRepo.persist(gm)
-    }
-
-    @Transactional
-    fun createSetting(id: UUID, gmId: UUID): Setting {
-        val gm = gmRepo.findById(gmId)
-        val setting = Setting().apply {
-            this.id = id
-            title = "world"
-            this.gm = gm
-        }
-        settingRepo.persist(setting)
-        return setting
-    }
-
-    @Transactional
-    fun createCampaign(id: UUID, gmId: UUID, setting: Setting): Campaign {
-        val gm = gmRepo.findById(gmId)
-        val camp = Campaign().apply {
-            this.id = id
-            title = "camp"
-            startedOn = Instant.now()
-            this.gm = gm
-            this.setting = setting
-        }
-        campaignRepo.persist(camp)
-        return camp
-    }
-
-    fun createCalendar(settingId: UUID): String {
-        val dto = CalendarDTO(null, "Cal", "CE", "[]", null, settingId)
-        return given()
-            .contentType(ContentType.JSON)
-            .body(dto)
-            .`when`().post("/api/settings/$settingId/calendars")
-            .then().statusCode(200)
-            .extract().path("id")
-    }
-
-    fun createEvent(calId: String, title: String, day: Int): String {
-        val dto = TimelineEventDTO(null, UUID.fromString(calId), title, null, day, null)
-        return given()
-            .contentType(ContentType.JSON)
-            .body(dto)
-            .`when`().post("/api/calendars/$calId/events")
-            .then().statusCode(200)
-            .extract().path("id")
     }
 
     @Test
