+++ src/test/kotlin/org/fg/ttrpg/TemplateResourceIT.kt
@@ -5,30 +5,14 @@ import io.quarkus.test.security.TestSecurity
 import io.quarkus.test.security.jwt.Claim
 import io.quarkus.test.security.jwt.JwtSecurity
 import io.restassured.RestAssured.given
-import jakarta.inject.Inject
-import jakarta.transaction.Transactional
-import org.fg.ttrpg.account.GM
-import org.fg.ttrpg.account.GMRepository
-import org.fg.ttrpg.setting.*
+import org.fg.ttrpg.testutils.IntegrationTestHelper
 import org.hamcrest.CoreMatchers.equalTo
-import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.util.*
 
 @QuarkusTest
-class TemplateResourceIT {
-    @Inject
-    lateinit var gmRepo: GMRepository
-
-    @Inject
-    lateinit var settingRepo: SettingRepository
-
-    @Inject
-    lateinit var templateRepo: TemplateRepository
-
-    @Inject
-    lateinit var genreRepo: org.fg.ttrpg.genre.GenreRepository
+class TemplateResourceIT : IntegrationTestHelper() {
 
     val gmId = UUID.fromString("00000000-0000-0000-0000-000000000001")
 
@@ -37,53 +21,6 @@ class TemplateResourceIT {
         createGm(gmId)
     }
 
-    @AfterEach
-    fun cleanup() {
-        gmRepo.deleteById(gmId)
-    }
-
-    @Transactional
-    fun createGm(id: UUID) {
-        val gm = GM().apply {
-            this.id = id
-            username = "gm-$id"
-        }
-        gmRepo.persist(gm)
-    }
-
-    @Transactional
-    fun createSetting(gmId: UUID): Pair<Setting, org.fg.ttrpg.genre.Genre> {
-        val gm = gmRepo.findById(gmId)
-        val setting = Setting().apply {
-            id = UUID.randomUUID()
-            title = "world"
-            this.gm = gm
-        }
-        settingRepo.persist(setting)
-        val genre = org.fg.ttrpg.genre.Genre().apply {
-            id = UUID.randomUUID()
-            title = "genre"
-            code = "code-${id.toString().substring(0,8)}"
-            this.setting = setting
-        }
-        genreRepo.insert(genre)
-        return setting to genre
-    }
-
-    @Transactional
-    fun createTemplate(genre: org.fg.ttrpg.genre.Genre, gmId: UUID, type: String): Template {
-        val gm = gmRepo.findById(gmId)
-        val template = Template().apply {
-            id = UUID.randomUUID()
-            title = "tpl"
-            this.type = type
-            jsonSchema = "{}"
-            this.genre = genre
-            this.gm = gm
-        }
-        templateRepo.persist(template)
-        return template
-    }
 
     @Test
     @TestSecurity(user = "userJwt", roles = ["viewer"])
@@ -95,12 +32,11 @@ class TemplateResourceIT {
         ]
     )
     fun listByGenre() {
-        val (setting, genre) = createSetting(gmId)
-        createTemplate(genre, gmId, "npc")
-        createTemplate(createSetting(gmId).second, gmId, "item")
+        val tpl = createTemplate(UUID.randomUUID(), gmId, "{}", "npc")
+        createTemplate(UUID.randomUUID(), gmId, "{}", "item")
 
         given()
-            .`when`().get("/api/templates?genre=${genre.id}")
+            .`when`().get("/api/templates?genre=${tpl.genre!!.id}")
             .then().statusCode(200)
             .body("size()", equalTo(1))
     }
@@ -115,9 +51,8 @@ class TemplateResourceIT {
         ]
     )
     fun listByType() {
-        val (setting, genre) = createSetting(gmId)
-        createTemplate(genre, gmId, "npc")
-        createTemplate(createSetting(gmId).second, gmId, "item")
+        createTemplate(UUID.randomUUID(), gmId, "{}", "npc")
+        createTemplate(UUID.randomUUID(), gmId, "{}", "item")
 
         given()
             .`when`().get("/api/templates?type=npc")
